refactor(frontend-demo): move Message type to lib/api
Centralize chat message shape with other API-related types; useChat and
MessageList import Message from api.

Made-with: Cursor
